+++ stix_shifter/stix_transmission/src/modules/cloudIdentity/cloudIdentity_ping.py
@@ -3,7 +3,7 @@ import json
 from .....utils.error_response import ErrorResponder
 
 
-class CloudIdentityPing(BaseQueryConnector):
+class CloudIdentityPing(BasePing):
     def __init__(self, api_client):
         self.api_client = api_client
